Adicionando o frontend

+++ Models/Turma.cs
@@ -7,7 +7,7 @@ public class Turma
     [Key]
     public long IdTurma { get; set; }
     public string Nome { get; set; }
-    public byte Ano { get; set; }
+    public short Ano { get; set; }
     public List<Aluno> Alunos { get; set; }
     public List<Professor> Professores { get; set; }
     public Escola Escola { get; set; }
